[add]添加广告投放业务层实现

+++ ad-service/ad-sponsor/src/main/java/com/glw/ad/service/vo/request/AdUnitRequest.java
@@ -20,7 +20,7 @@ public class AdUnitRequest {
 
     private String unitName;
 
-    private Integer positionType;
+    private Byte positionType;
 
     private Long budget;
